+++ CHANGELOG.md
@@ -26,13 +26,13 @@ This document details all notable changes to the WSU HRS Child Theme. Uses [Sema
 
 - Set max width on text-based content like posts and standard pages, open #30.
 - Set up grid fallbacks, open #31.
-- Set up no-js handling for the search form and menu, open #34.
 - Get Babel processing working, open #35.
 
 ## 0.15.2 (unreleased)
 
 ### Fixed
 
+- Add noscript element to handle no-JS search menu display to fix #34.
 - Add focus visibility to the search menu close button to fix #33.
 
 ### Changed